+++ index.html
@@ -228,6 +228,6 @@
     </div>
   </template>
 
-  <script src="js/main.js"></script>
+  <script src="js/main.js" type="module"></script>
 </body>
 </html>
